Move sku-related functions to `skuVariations.ts`
